tyckers: introduce `ConcreteAwareTycker` so that `StatedTycker` can be a superclass
